requirements.txt
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
asttokens==2.4.1
attrs==24.2.0
billiard==4.2.0
black==24.8.0
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
executing==2.1.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
importlib_resources==6.4.4
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.2
sentry-sdk==2.13.0
six==1.16.0
SQLAlchemy==2.0.32
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
